Move constants to seperate files and update preview functions names
