+++ Chrono/Chrono/util/constants/Strings.swift
@@ -1,0 +1,128 @@
+//
+//  Strings.swift
+//  Chrono
+//
+//  Created by Hari Govind on 2021-07-07.
+//
+
+import Foundation
+struct Strings{
+    static let appName = "Chrono"
+    
+    // Timer Controls
+    static let stop = "Stop"
+    static let start = "Start"
+    static let pause = "Pause"
+    static let resume = "Resume"
+    static let lap = "Lap"
+    static let chronometer = "Chronometer"
+    
+    // Circuit Create
+    static let _0 = "0"
+    static let _5 = "5"
+    static let _30 = "30"
+    static let nameInputHint = "Name of Circuit"
+    static let save = "Save"
+    static let discard = "Discard"
+    static let todo = "TODO"
+    static let placeholder = "Placeholder"
+    static let sets = "Sets"
+    static let dismiss = "Dismiss"
+    static let selectAnIcon = "Select an Icon"
+    static let sec = "sec"
+    static let rest = "Rest"
+    static let work = "Work"
+    
+    static let workout = "Workout"
+    static let getReady = "Get Ready"
+    static let circuitName = "Circuit Name"
+    static let name = "Name"
+    static let letsGo = "Let's Go!"
+    
+    // Tab Navigation
+    static let settings = "Settings"
+    static let circuits = "Circuits"
+    static let stopwatch = "Stopwatch"
+    
+    // Delete Circuit Dialog
+    static let delete_cicuit = "Delete Circuit"
+    static let delete_circuit_subtitle = "Do you want to delete this circuit?"
+    static let delete = "Delete"
+    static let cancel = "Cancel"
+    
+    // Circuit Complete Dialog
+    static let circuit_complete = "Circuit Complete🥳🥳🥳"
+    static let circuit_complete_subtitle = "Woohoo! You have completed your circuit, good job!!!"
+    static let circuit_complete_confirm = "Finish"
+    static let circuit_complete_cancel = "Restart"
+    static let circuit_complete_celebrate = "Circuit Complete!"
+    
+    // Easter Egg Dialog
+    static let easter_egg = "🐇🥚"
+    static let easter_egg_subtitle = "Congrats! You just unlocked the super secret stats page. Soon, this will be available to everyone, but right now it\'s an easter egg 🤫"
+    static let easter_egg_confirm = "Continue"
+    static let stats = "Stats"
+    static let total_time = "Total Time"
+    static let total_circuits = "Total Circuits"
+    static let _0h0m = "0h 0m"
+    static let _0h0m0s = "0h 0m 0s"
+    
+    
+    // Settings
+    static let notifications = "Notifications"
+    static let on = "On"
+    static let dark_mode = "Dark Mode"
+    static let off = "Off"
+    static let circuit_timer = "Circuit Timer"
+    static let preferences = "Preferences"
+    static let audio_prompts = "Audio Pompts"
+    static let about = "About"
+    static let set = "Set"
+    static let privacy_policy = "Privacy Policy"
+    static let get_ready_time = "Get Ready Time"
+    static let time_emoji = "⏳"
+    static let dark_emoji = "🕶️"
+    static let _5s = "5s"
+    static let _10s = "10s"
+    static let _15s = "15s"
+    static let rate_app = "Rate App"
+    static let get_help = "Get Help"
+    static let skip_last_rest = "Skip Last Rest"
+    static let select_icon_subtitle = "Use this icon to represent your circuit"
+    
+    static let chrono_logo = "Chrono Logo"
+    static let change_nicon = "Change\nIcon"
+    static let max_char_warning = "Maximum 50 characters."
+    static let coming_soon = "Coming Soon"
+    static let system_default = "System Default"
+    static let the_number_of_seconds_before_your_circuit_begins = "The number of seconds before your circuit begins"
+    static let enter_the_dark_side = "Enter the dark side"
+    static let edit = "Edit"
+    static let share = "Share"
+    static let update_available_title = "Update Available 📲"
+    static let update_available_subtitle = "You are on an older version of Chrono. Please update your app for the latest features and bug fixes."
+    static let update = "Update"
+    static let help_email_subject = "Chrono Support Ticket"
+    static let help_email_body = "Please let us know what you need help with. Feel free to include screenshots:"
+    static let sound_effect = "Sound Effect"
+    static let speaker_emoji = "🔊"
+    static let sound_effect_subtitle = "Pick the sound you like for your circuits"
+    static let short_whistle = "Short Whistle"
+    static let long_whistle = "Long Whistle"
+    static let next = "Next"
+    static let skip = "Skip"
+    static let continue_button = "Continue"
+    
+    // Onboarding Titles/Subtitles
+    static let onboard_welcome_title = "Welcome"
+    static let onboard_welcome_subtitle = "Thank you for choosing Chrono! Let\'s\nwalk you through how our app\nworks."
+    static let onboard_circuit_title = "Save custom circuits"
+    static let onboard_circuit_subtitle = "Customize your circuit with your\n desired times, title, and icon. Click on\n your saved circuit to run it."
+    static let onboard_swatch_title = "Use the stopwatch"
+    static let onboard_swatch_subtitle = "With integrated lap functionality, our\n stopwatch will help you time\n basically anything."
+    static let onboard_setting_title = "Dark mode"
+    static let onboard_setting_subtitle = "Use the settings page to personalize\nyour Chrono experience. Also there\'s\n dark mode. 🖐 🎤 ⬇"
+    static let onboard_final_title = "You\'re all set"
+    static let onboard_final_subtitle = "Feel free to leave us feedback and\" joining our Discord server. We\n hope you enjoy using Chrono!"
+    static let discord = "Discord"
+}
