+++ Chrono/Chrono/util/constants/links.swift
@@ -1,0 +1,14 @@
+//
+//  links.swift
+//  Chrono
+//
+//  Created by Hari Govind on 2021-07-07.
+//
+
+import Foundation
+struct links{
+    static let privacy_page_link = "https://chronofit.herokuapp.com/privacyPolicy"
+    static let discord_invite_link = "https://discord.gg/UbWrRxwkgR"
+    static let app_rating_primary_link = "TODO"
+    static let app_rating_secondary_link = "TODO"
+}
